# shared/api/order_v1.swagger.json
{
  "info": {
    "description": "API для работы с заказами",
    "title": "OrderService API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/orders": {
      "post": {
        "operationId": "CreateOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "part_uuids": [
                    "111e4567-e89b-12d3-a456-426614174001",
                    "222e4567-e89b-12d3-a456-426614174002"
                  ],
                  "user_uuid": "123e4567-e89b-12d3-a456-426614174000"
                },
                "properties": {
                  "part_uuids": {
                    "description": "Массив UUID выбранных деталей",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "user_uuid": {
                    "description": "UUID пользователя, создающего заказ",
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "user_uuid",
                  "part_uuids"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "order_uuid": "333e4567-e89b-12d3-a456-426614174003",
                    "total_price": 123.45
                  },
                  "properties": {
                    "order_uuid": {
                      "description": "UUID созданного заказа",
                      "format": "uuid",
                      "type": "string"
                    },
                    "total_price": {
                      "description": "Итоговая стоимость заказа",
                      "format": "float",
                      "type": "number"
                    }
                  },
                  "required": [
                    "order_uuid",
                    "total_price"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Orders info"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 400,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Bad Request: Invalid parameter format",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 500,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Internal server error occurred",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "summary": "Создание заказа",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{order_uuid}": {
      "get": {
        "operationId": "GetOrder",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "order_uuid": "333e4567-e89b-12d3-a456-426614174003",
                    "part_uuids": [
                      "111e4567-e89b-12d3-a456-426614174001",
                      "222e4567-e89b-12d3-a456-426614174002"
                    ],
                    "payment_method": "CARD",
                    "status": "PAID",
                    "total_price": 123.45,
                    "transaction_uuid": "444e4567-e89b-12d3-a456-426614174004",
                    "user_uuid": "123e4567-e89b-12d3-a456-426614174000"
                  },
                  "properties": {
                    "order_uuid": {
                      "description": "Уникальный идентификатор заказа",
                      "format": "uuid",
                      "type": "string"
                    },
                    "part_uuids": {
                      "description": "Список UUID деталей",
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "payment_method": {
                      "description": "Способ оплаты",
                      "enum": [
                        "UNKNOWN",
                        "CARD",
                        "SBP",
                        "CREDIT_CARD",
                        "INVESTOR_MONEY"
                      ],
                      "nullable": true,
                      "type": "string",
                      "x-enumDescriptions": {
                        "CARD": "Банковская карта",
                        "CREDIT_CARD": "Кредитная карта",
                        "INVESTOR_MONEY": "Деньги инвестора",
                        "SBP": "Система быстрых платежей",
                        "UNKNOWN": "Неизвестный способ"
                      }
                    },
                    "status": {
                      "description": "Статус заказа",
                      "enum": [
                        "PENDING_PAYMENT",
                        "PAID",
                        "CANCELLED"
                      ],
                      "example": "PENDING_PAYMENT",
                      "type": "string"
                    },
                    "total_price": {
                      "description": "Итоговая стоимость",
                      "format": "double",
                      "type": "number"
                    },
                    "transaction_uuid": {
                      "description": "UUID транзакции (если оплачен)",
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "user_uuid": {
                      "description": "UUID пользователя",
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "order_uuid",
                    "user_uuid",
                    "part_uuids",
                    "total_price",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Order info"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 404,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Order with UUID '123' not found",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Order not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 500,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Internal server error occurred",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "summary": "Получение заказа",
        "tags": [
          "Orders"
        ]
      },
      "parameters": [
        {
          "description": "UUID заказа",
          "in": "path",
          "name": "order_uuid",
          "required": true,
          "schema": {
            "example": "fabc28ad-c834-4a83-8f2a-67543c1ab0b0",
            "format": "uuid",
            "type": "string"
          }
        }
      ]
    },
    "/api/v1/orders/{order_uuid}/cancel": {
      "parameters": [
        {
          "description": "UUID заказа",
          "in": "path",
          "name": "order_uuid",
          "required": true,
          "schema": {
            "example": "fabc28ad-c834-4a83-8f2a-67543c1ab0b0",
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "CancelOrder",
        "responses": {
          "204": {
            "description": "Order canceled"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 400,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Bad Request: Invalid parameter format",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad request - validation error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 404,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Order with UUID '123' not found",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Order not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 409,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Order cannot be canceled",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Cannot cancel order, it is already PAID."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 500,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Internal server error occurred",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "summary": "Отмена заказа",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{order_uuid}/pay": {
      "parameters": [
        {
          "description": "UUID заказа",
          "in": "path",
          "name": "order_uuid",
          "required": true,
          "schema": {
            "example": "fabc28ad-c834-4a83-8f2a-67543c1ab0b0",
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "PayOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "payment_method": "CARD"
                },
                "properties": {
                  "payment_method": {
                    "description": "Метод оплаты",
                    "enum": [
                      "CARD",
                      "SBP",
                      "CREDIT_CARD",
                      "INVESTOR_MONEY"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "payment_method"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "transaction_uuid": "333e4567-e89b-12d3-a456-426614174003"
                  },
                  "properties": {
                    "transaction_uuid": {
                      "description": "UUID транзакции оплаты",
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "transaction_uuid"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Order created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 400,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Bad Request: Invalid parameter format",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 404,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Order with UUID '123' not found",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Order not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "example": 500,
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "example": "Internal server error occurred",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "description": "HTTP-код ошибки",
                      "type": "integer"
                    },
                    "message": {
                      "description": "Описание ошибки",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "summary": "Оплата заказа",
        "tags": [
          "Orders"
        ]
      }
    }
  },
...
}